+++ src/hooks/login/index.js
@@ -1,6 +1,5 @@
 import useFacebookLogin from './useFacebookLogin';
-import useGoogleLogin from './useGoogleLogin';
 import useLogin from './useLogin';
 import useLogout from './useLogout';
 
-export { useFacebookLogin, useGoogleLogin, useLogin, useLogout };
+export { useFacebookLogin, useLogin, useLogout };
